pyproject.toml
[tool.poetry]
name = "vail-sphere-rest-api-client"
version = "2.0"
description = "A client library for accessing Vail Sphere REST API"
authors = []
readme = "README.md"
packages = [
    {include = "vail_sphere_rest_api_client"},
]
include = ["CHANGELOG.md", "vail_sphere_rest_api_client/py.typed"]


[tool.poetry.dependencies]
python = "^3.8"
httpx = ">=0.20.0,<0.27.0"
attrs = ">=21.3.0"
python-dateutil = "^2.8.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
select = ["F", "I", "UP"]
line-length = 120
